requirements.txt
amqp==2.6.0
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.1.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.6.3.0
boto3==1.7.75
botocore==1.10.75
cached-property==1.4.3
celery==4.3.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
confluent-kafka==1.0.1
cryptography==2.3
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.11
django-choices==1.6.0
django-filter==2.0.0
django-nose==1.4.5
django-partial-index==0.5.2
django-reversion==2.0.13
django-reversion-compare==0.8.6
djangorestframework==3.11.0
docutils==0.14
Fabric3==1.14.post1
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
importlib-metadata==1.7.0
ipdb==0.13.8
ipython==7.24.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
jedi==0.18.0
jmespath==0.9.3
kazoo==2.5.0
kombu==4.6.11
libfaketime==1.2.1
lxml==4.2.3
Markdown==2.6.11
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy==0.610
newrelic==3.2.2.94
nose==1.3.7
paramiko==2.4.2
parso==0.8.2
pexpect==4.6.0
phonenumbers==8.11.5
pickleshare==0.7.4
pkcs7==0.1.2
prompt-toolkit==3.0.18
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.9.4
pyflakes==1.6.0
Pygments==2.2.0
pykafka==2.8.0
PyNaCl==1.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.4.1
redis==3.3.11
requests==2.20.0
requests-mock==1.5.0
requests-toolbelt==0.8.0
s3transfer==0.1.13
sentry-sdk==0.8.0
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
tabulate==0.8.2
toml==0.10.2
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
vine==1.3.0
wcwidth==0.1.7
whitenoise==4.0
zeep==3.0.0
zipp==3.1.0
